Moved Microphone dropdown into MicrophoneRecord (+other)
