+++ Packages/com.whisper.unity/Runtime/Utils/MicrophoneRecord.cs.meta
